// src/schemas/authSchema.ts
import { z } from "zod";

export const signInSchema = z.object({
    email: z.string().email({ message: "Invalid email address" }),
    password: z.string().min(6, { message: "Password must be at least 6 characters" }),
});

export const signUpSchema = z.object({
    firstName: z.string().min(2, { message: "First name must be at least 2 characters" }),
    lastName: z.string().min(2, { message: "Last name must be at least 2 characters" }).optional().or(z.literal('')),
    email: z.string().email({ message: "Invalid email address" }),
    password: z.string().min(6, { message: "Password must be at least 6 characters" }),
});

export const verifySchema = z.object({
    email: z.string().email({ message: "Invalid email address" }),
    code: z.string().length(6, { message: "Verification code must be at least 6 digits" })
});